+++ accountresolver/resolver.go
@@ -12,7 +12,7 @@ import (
 )
 
 type AccountsResolver interface {
-	Extend(ctx context.Context, blockNum uint64, trxHash []byte, instructionIndex int, key Account, accounts Accounts) error
+	Extend(ctx context.Context, blockNum uint64, trxHash []byte, instructionIndex uint64, key Account, accounts Accounts) error
 	Resolve(ctx context.Context, atBlockNum uint64, key Account) (Accounts, bool, error)
 	StoreCursor(ctx context.Context, readerName string, cursor *Cursor) error
 	GetCursor(ctx context.Context, readerName string) (*Cursor, error)
@@ -37,7 +37,7 @@ func NewKVDBAccountsResolver(store store.KVStore, logger *zap.Logger) *KVDBAccou
 	}
 }
 
-func (r *KVDBAccountsResolver) Extend(ctx context.Context, blockNum uint64, trxHash []byte, instructionIndex int, key Account, accounts Accounts) error {
+func (r *KVDBAccountsResolver) Extend(ctx context.Context, blockNum uint64, trxHash []byte, instructionIndex uint64, key Account, accounts Accounts) error {
 	if !r.isKnownInstruction(ctx, trxHash, instructionIndex) {
 		return nil
 	}
@@ -121,7 +121,7 @@ func (r *KVDBAccountsResolver) Resolve(ctx context.Context, atBlockNum uint64, k
 	return resolvedAccounts, false, nil
 }
 
-func (r *KVDBAccountsResolver) isKnownInstruction(ctx context.Context, transactionHash []byte, instructionIndex int) bool {
+func (r *KVDBAccountsResolver) isKnownInstruction(ctx context.Context, transactionHash []byte, instructionIndex uint64) bool {
 	instructionKey := Keys.knownInstruction(transactionHash, instructionIndex)
 	_, err := r.store.Get(ctx, instructionKey)
 	return errors.Is(err, store.ErrNotFound)
